TT-1682: Add paused registration page.
IDPs are now able to send a PENDING response to the hub to indicate that
the user has paused their registration journey and intends to continue later.

A new page has been added to the frontend to handle this response.
The purpose of this page is to give the user instructions on how to resume
registration.

Authors: @michaelwalker, @mahmudh